Update code in homo lr
